requirements.txt
cffi
cloudpickle
cycler
dask
decorator
imageio
kiwisolver
matplotlib
networkx
numpy
pandas
Pillow
pycparser
pygit
pyparsing
python-dateutil
pytz
PyWavelets
PyYAML
scikit-learn
scipy
six
toolz
opencv-python
CMake
scikit-image
tqdm
